CoinApp/requirements.txt
django >= 2
djangorestframework
celery
dj-rest-auth
#django-celery